Refactor logging and update import conventions
Moved logging.ts to utils/logging.ts and updated all relevant imports to use the new path. Standardized import statements to use double quotes for consistency across config, server, and mutex modules.
