+++ server/src/config/config.ts
@@ -1,4 +1,4 @@
-import dotenv from 'dotenv';
+import dotenv from "dotenv";
 
 dotenv.config();
 
@@ -12,9 +12,9 @@ interface Config {
 
 const config: Config = {
   port: Number(process.env.PORT) || 3000,
-  nodeEnv: process.env.NODE_ENV || 'development',
-  dbUrl: process.env.DB_URL || '',
-  dbName: process.env.DB_NAME || '',
+  nodeEnv: process.env.NODE_ENV || "development",
+  dbUrl: process.env.DB_URL || "",
+  dbName: process.env.DB_NAME || "",
   dbConnectionRetryCount: Number(process.env.DB_CONNECTION_RETRY_COUNT) || 5
 };
